Move Error module into utils/
